# product-api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8083","description":"Generated server url"}],"tags":[{"name":"Product management","description":"API for product management"}],"paths":{"/api/v1/products/{id}":{"get":{"tags":["Product management"],"summary":"Get product by id","description":"Retrieve detailed information about a specific product","operationId":"getProductById","parameters":[{"name":"id","in":"path","description":"Product ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"put":{"tags":["Product management"],"summary":"Update Product","description":"Update product only, not image","operationId":"updateProduct","parameters":[{"name":"id","in":"path","description":"Product ID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"delete":{"tags":["Product management"],"summary":"Delete a product","description":"Delete a product by id and image in cloudinary","operationId":"deleteProductById","parameters":[{"name":"id","in":"path","description":"Product ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/products/{id}/image/order":{"put":{"tags":["Product management"],"summary":"Update image order","description":"Update image order of product, images must belong to the product","operationId":"updateProductImageOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"string","format":"uuid"}}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/v1/products":{"post":{"tags":["Product management"],"summary":"Create Product","description":"APi for create product, call API cloudinary before create to upload image","operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/v1/products/{productId}/image":{"post":{"tags":["Product management"],"summary":"Add a single image to a product","description":"Add one image to the end of image list","operationId":"addProductImage","parameters":[{"name":"productId","in":"path","description":"Product ID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"productImages","in":"query","description":"Image file","required":true,"schema":{"type":"array","items":{"type":"string","format":"binary"}}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/v1/comments/product/{productId}":{"get":{"tags":["comment-controller"],"operationId":"findByProductId","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCommentResponse"}}}}}},"post":{"tags":["comment-controller"],"operationId":"createComment","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommentRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentResponse"}}}}}}},"/api/v1/products/search":{"get":{"tags":["Product management"],"summary":"Search product","description":"Search product by name and category and filter by price","operationId":"getProducts","parameters":[{"name":"name","in":"query","description":"Product name","required":false,"schema":{"type":"string"}},{"name":"categoryId","in":"query","description":"CategoryId","required":false,"schema":{"type":"string","format":"uuid"}},{"name":"pageable","in":"query","description":"pageable: ?page= &size= &sort= ,asc/desc","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageProductResponse"}}}}}}},"/api/v1/comments/user/{userId}":{"get":{"tags":["comment-controller"],"operationId":"findByUserId","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCommentResponse"}}}}}}},"/api/v1/comments/product/{productId}/rating/{rating}":{"get":{"tags":["comment-controller"],"operationId":"findByProductIdAndRating","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"rating","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCommentResponse"}}}}}}},"/api/v1/products/{id}/image/{imageId}":{"delete":{"tags":["Product management"],"summary":"Delete an image","description":"Delete an image of a product","operationId":"deleteProductImage","parameters":[{"name":"id","in":"path","description":"Product ID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"imageId","in":"path","description":"Image ID (Not publicId)","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"ApiError":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"UpdateProductRequest":{"type":"object","description":"Update product request","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32"},"categoryId":{"type":"string","format":"uuid"}}},"ProductImage":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"imageUrl":{"type":"string","minLength":1},"publicId":{"type":"string"}},"required":["imageUrl"]},"ProductResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"description":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"price":{"type":"integer","format":"int32"},"categoryId":{"type":"string","format":"uuid"},"categoryName":{"type":"string"},"productImages":{"type":"array","items":{"$ref":"#/components/schemas/ProductImage"}}}},"CreateProductRequest":{"type":"object","description":"CreateProductRequest","properties":{"name":{"type":"string","minLength":1},"description":{"type":"string"},"price":{"type":"integer","format":"int32"},"quantity":{"type":"integer","format":"int32","minimum":0},"categoryId":{"type":"string","format":"uuid"},"categoryName":{"type":"string"},"productImages":{"type":"array","items":{"$ref":"#/components/schemas/ImageRequest"}}},"required":["name","price","quantity"]},"ImageRequest":{"type":"object","properties":{"imageUrl":{"type":"string","minLength":1},"publicId":{"type":"string","minLength":1}},"required":["imageUrl","publicId"]},"CreateCommentRequest":{"type":"object","properties":{"comment":{"type":"string","minLength":1},"rating":{"type":"integer","format":"int32","maximum":5,"minimum":0},"productId":{"type":"string","format":"uuid"}},"required":["comment","productId"]},"CommentResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"comment":{"type":"string"},"rating":{"type":"integer","format":"int32"},"userName":{"type":"string"},"avatarUrl":{"type":"string"},"createdDate":{"type":"string","format":"date-time"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageProductResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"sorted":{"type":"boolean"},"empty":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"PageCommentResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/CommentResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}}}}}